RE-1191 Add common.globalwraps
This function is called to wrap every job.
This allows us to add and remove global wrappers in a single location.
Currently the global wrappers are:
  * timestamps
  * timeout (10 hours), jobs can still set their own shorter timeouts,
    This timeout was chosen to be longer than any existing timeout.
  * Logfile size checker: Aborts build if console log exceeds 200MB.
  * common.shared_slave: Logfile size checker requires a node to
    execute on. This means that jobs that use globalwraps (all jobs)
    don't need to call common.shared_slave separately unless they need
    multiple nodes.

Jobs that previously used common.shared_slave have had that call
replaced with common.globalWraps, jobs that used other types of slave
have been wrapped in a common.globalWraps call.